Test triggering Circle CI workflow with git commit
Perhaps the issue with the site not seemingly pulling new content in from Contentful (despite Circle CI being triggered and successful "build" being run upon Contentful changes) is that the Circle CI "workflow" pipeline is not triggered. In the past it has been triggered by a git commit to the remote repository on GitHub. The problem could lie in changing the settings to trigger the workflow (or pushing a GitHub commit to trigger the workflow) upon Contentful updates.